+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@fgv/ts-json",
-  "version": "1.9.4",
+  "version": "1.9.5",
   "description": "Typescript utilities for working with JSON",
   "main": "dist/index.js",
   "scripts": {
@@ -25,22 +25,22 @@
   "author": "Erik Fortune",
   "license": "MIT",
   "devDependencies": {
-    "@fgv/ts-utils": "^1.9.3",
+    "@fgv/ts-utils": "^1.9.4",
     "@fgv/ts-utils-jest": "^1.9.4",
-    "@types/jest": "^29.4.0",
+    "@types/jest": "^29.5.0",
     "@types/mustache": "^4.2.2",
-    "@types/node": "^18.14.6",
-    "@typescript-eslint/eslint-plugin": "^5.54.0",
-    "@typescript-eslint/parser": "^5.54.0",
-    "eslint": "^8.35.0",
+    "@types/node": "^18.15.10",
+    "@typescript-eslint/eslint-plugin": "^5.56.0",
+    "@typescript-eslint/parser": "^5.56.0",
+    "eslint": "^8.36.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
-    "jest": "^29.4.3",
+    "jest": "^29.5.0",
     "jest-extended": "^3.2.4",
     "mustache": "^4.2.0",
-    "rimraf": "^4.3.0",
+    "rimraf": "^4.4.1",
     "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5"
